ASG3: Removed overloaded methods to be used in the UI;
